Relatorio_valorant/src/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
python-dotenv
requests
beautifulsoup4